// Source/DocuEye.Web/client-app-old/src/api/docueye-api/models/filtered-view.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DocuEye.Web
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { AutomaticLayout } from './automatic-layout';
// May contain unused imports in some cases
// @ts-ignore
import { ElementView } from './element-view';
// May contain unused imports in some cases
// @ts-ignore
import { RelationshipView } from './relationship-view';

/**
 * 
 * @export
 * @interface FilteredView
 */
export interface FilteredView {
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'id'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'description'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'key'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'workspaceId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'viewType'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'paperSize'?: string | null;
    /**
     * 
     * @type {Array<ElementView>}
     * @memberof FilteredView
     */
    'elements'?: Array<ElementView> | null;
    /**
     * 
     * @type {Array<RelationshipView>}
     * @memberof FilteredView
     */
    'relationships'?: Array<RelationshipView> | null;
    /**
     * 
     * @type {AutomaticLayout}
     * @memberof FilteredView
     */
    'automaticLayout'?: AutomaticLayout;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'baseViewKey'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FilteredView
     */
    'mode'?: string | null;
    /**
     * 
     * @type {Array<string>}
     * @memberof FilteredView
     */
    'tags'?: Array<string> | null;
}